refactor: Move RunState to the ImportAutomation namespace
